+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org) using the following scheme: MAJOR.MINOR.PATCH
 
+## [0.6.0] - 2026-03-16
+
+### Changed
+- Updated DolibarrObjectStatus color property to foregroundStyle
+- Updated swift tools version to 6.1
+
 ## [0.5.0] - 2026-03-15
 
 ### Added